Fixed some call back parameter bugs
Callback parameters  were incorrect for two reasons.
The first is the shift of the first half of the 64-bit parameter was
only 16 bits in a number of areas.
The second is that a C# casting problem was ANDing the two 32-bits into
one 32-bit instead of creating a 64-bit parameter.