C# callback problem fixed
If a while(true){} is in the main routine for C#,this no longer breaks
our code.
This was happening because the code compiles into infinite branches,and
the advanced timer rewrite broke the timeouts in the interpreter.cpp.
This is working now, however the HAL_COMPLETION timeouts do not timeout
exactly when you want, because for some reason enabling the Count
Comparater 2 on TIM1 and TIM2 breaks the COM port completely.
The timeouts do work well enough though but should be fixed to exact
time.