Radio fix.
The problem was that when trying to send the beacon and the data packet that if one didn't go through properly then the other ended up staying in the send queue.
The fix was to add a send flush timer where if more than one packet was in the send queue when attempting to send a packet, the timer would be turned on and the extra packets would eventually all be flushed out. Once the packets were flushed out the timer was turned off to keep load down.