Fixed deployment problems.
Removed some previous fixes that were no longer needed.
Cleaned up the code and commented the fixes.

The root problem is that after we erase the deployment area, managed threads can still be accessed by the Garbage Collector and the Execution engine. Compaction also was causing problems after the Garbage Collection.

During deployment we erase the Flash and then load in new code which is then executed from the beginning. So our fix here is to set a flag, after VS tells us to erase the Flash, which keeps us from executing managed threads, or Mark() user threads, and we stop compaction after Garbage Collection. If we get a Ping packet or restart we reset the flag and execute everything normally.