.NOW UART: Only use managed queues when one exists
By default the serial code was writing to both managed and unmanaged
queues for COM1 which was doubling a portion of the work.

This is only needed if we have a managed serial port in C#. This often
isn't the case. So we check for that, and if there isn't one, we can skip
the asscioated lock and write.

Otherwise we were doing too much work. This fixes #249.

This helps the average case especially for clean deployment. However the
worst case performance will still apply when a serial port on COM1 is
declared from C#.