* Spring Camp Only ADC and bootloader changes.
The ADC driver peeks at the MSM interrupt pin after each buffer. This is
strange, normally this is an output not an input, but we are hard up
because of a hardware issue on the hard reset line. If this peek reads 1,
it will reboot the node. Because the peek input is also the BOOT0 pin,
this will bring us up in the hardware bootloader. From there we can use
stm32flash from Android.