For radars with new mux enable and 20ms guard time.
Modified radars have new pin to control mux. This allows us to silence the
radar without blowing up the filters, which happens if the mux is just
left on.

A 20ms guard time, meaning the radar pulses are running for 20ms before
the mux is turned on, is used here. Even with this, the radar IQ signals
don't fully recover until 10ms AFTER the mux is turned back on.

So the "lost" time is pulse_off + 20ms + 10ms. A typical pulse_off is 3ms
or so, although it can vary between 1ms and about 20ms.