Use interupts for TIM7. Make interrupts atomic.
Used interrupts on TIM7 to avoid busy-waits which might interfere with
ADC.

Added checks and made interrupts atomic to make sure there won't be
problems with ACKs, which is a second off state that comes very quickly
after an TX event. So we double-check that we aren't in an ACK state
before we turn on the radar.