Adds (hacks) very basic support for TinyOS apps
All currently disabled (mostly ifdef'd out) by default.

This simply checks a set memory location for a set key. If it finds the
key, it will attempt to run whatever is at 0x08020000 (same as TinyCLR).

There are corrosponding changes to TinyOS to match this (adds keys,
changes binary base, tweaks to startup code, etc.).

I did not augment the "Engine" scheme in TinyBooter. If it finds the key,
it goes. Note that it still honors the boot pin allowing for deploying a
new app. It is possible for a TinyOS app to implement the "Stop And
Reboot" function allowing for a software reboot to bootloader, but there
is no mandate for that currently.

No changes were needed to programming. The binary is not checked (or has
already been disabled). TinyBooter will happily program any SREC you give
it, provided the addresses match up (here 0x08020000).

I also removed some excess text from the startup sequence. As a result
TinyBooter will not output anything over serial unless it actually invoked
to reprogram. So a TinyOS app will be "clean" and not output anything that
isn't in the app itself.