Changes to fix issue where DataTransmitter throws error that broadcast address was not found
During initial startup, discovery messages are sent to all neighbors
(0xffff). This entry (65535) is also added to the neighbor table. When
there is a discovery (a msg from another node), this entry is added to the
neighbor table. After sometime (when system stabilizes - there is disco
msgs back and forth), when there are no more msgs from 65535,
DataTransmitter complains that there is no neighbor called 65535. This is
one of the fixes of this commit.

Individual header files are no longer included in OMAC's proj file.

There are a few style fixes.