Able to call the app's receive function from MAC. But there are still crashes and the design needs to be changed.
Disco and DataSend make use of a global buffer to send and receive data
and call the radio's send directly, whereas timesync does not use a global
buffer and calls the MAC's send function. MAC's send function, in an
earlier version, fills up the global buffer and calls radio's send.