.NOW: Re-init the USART when changing power levels.
The UART module needs to be clued-in to the clock speed.
So when we change it from the power driver, we must re-init USART.

Note that this will likely cause problems with any USART transactions in
progress. For now, I insert a Thread.Sleep(100) to ensure buffers are
flushed before the PowerState command. TODO: Move this step down a layer.