.NOW UART init re-org.
This is to prepare for power driver architecture improvements.

UART has a special relationship with the power driver because it sits in
the usb-power path, and because its critical for user interaction via
MFDeploy or VS and we so must be very deliberate with it.

The UART driver is now aware enough to know when usb-serial is actually
plugged in. I needed this for power management purposes and was the real
goal. As a side effect, COM0 will not actually init the UART until
something is attached, which is nice.

Only plug events are registered. No un-plug yet. May not add it.

The new plug-in logic does not apply to COM1 for obvious reasons.