Fix unknown-pragmas check_printf_formats
Part of -Wall clean-up for better code. See #234

check_printf_formats is an ARM-compiler specific pragma and not in GCC.
However there is equivilent functionality in GCC with function attribute
'format'. I have made the substitution.

This gives the compiler a hint to allow for better compile time checking
of such functions.

These changes will properly eliminate the warning and improve compile time
error checking, but is not expected to change the binary.