Rework of C# performance benchmark suite.
Garbage collect in C# doesn't respect bracket scope, apparently. Was
losing memory between tests and the GC wasn't getting it back.

So I either had to manually null everything... or just make it a static
function.

Also the old code was ugly. This isn't as pretty and clever as it could
be. But it's a start.