Adds a new virtual timer to DataTxHandler
The new virtual timer separates DataTx from the scheduler's timer.

This change also removes the additional GUARDBAND time that was added to
DataTx in scheduler. This along with an earlier allows DataTx to wake up
one GUARDTIME period early, listen for other transmitters and transmit
only when channel is clear (CCA).