Retransmit current packet if there is still sufficient time left in current rx slot
If send did not succeed for some reason (tx woke up earlier than rx), then
resend if there is still time left in rx slot.
This logic is not good enough for the case where tx wakes up after rx slot
and tries to send when rx is not listening at all.