.NOW: Minor. Move const array to rodata
STM RCC code uses 'volatile const' for some tables. Despite being 'const',
the 'volatile' forces GCC allocate space in rwdata (i.e. SRAM).

As I cannot fathom any reason for such a table to be 'const volatile'
(this is typically only used for memory mapped IO), we can drop the
'volatile'.

Frees up a little RAM, 20 bytes. Should glance over rest of code to see if
anything else is out there.

I suspect that on other compilers the use of 'volatile' in this case is
harmless and/or its a bug.