Post SONYC NYC Demo SX1276 fixes.
--PACONFIG register is never touched by this driver and the default
setting is completely wrong which causes the driver to never use the
antenna. I have hacked in a default setting of +2dBm (including the use
of the antenna).

--The driver has a pre-exsisting bug that it ignores requests to change
the power setting. This remains.

--RadioRegsInit were not being spewed properly to the radio at startup.
The const was not getting populated or something. Thinking this is due
to bad C++, hopefully not a compile problem.

--Radio was not being reset on init. Fixed.